Fix export/import issues with Expandable.
