+++ src/reusable/expandable/expandable-children.js
@@ -1,15 +1,10 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { ExpandableContext } from './expandable'
-
 
 class ExpandableChildren extends Component {
   componentDidMount() {
-    const {
-      context,
-      children,
-      show
-    } = this.props;
+    const { context, children, show } = this.props
 
     if (children.length <= show && !context.disabled) {
       context.disable()
@@ -17,34 +12,26 @@ class ExpandableChildren extends Component {
   }
 
   render() {
-    const {
-      context,
-      children,
-      show
-    } = this.props;
-    
+    const { context, children, show } = this.props
+
     return (
       <React.Fragment>
-        {context.expanded ? (
-          children
-        ) : (
-          children.slice(0, show)
-        )}
+        {context.expanded ? children : children.slice(0, show)}
       </React.Fragment>
     )
   }
 }
 
 ExpandableChildren.propTypes = {
-  show: PropTypes.number
-};
+  show: PropTypes.number,
+}
 
 ExpandableChildren.defaultProps = {
-  show: 3
-};
+  show: 3,
+}
 
 export default props => (
   <ExpandableContext.Consumer>
-    {context => <ExpandableChildren {...props} context={context}/>}
+    {context => <ExpandableChildren {...props} context={context} />}
   </ExpandableContext.Consumer>
 )
