+++ src/reusable/expandable/expandable.js
@@ -1,25 +1,26 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 
-const ExpandableContext = React.createContext();
+const ExpandableContext = React.createContext()
 
 /**
-  Use Expandable to show only the first few items. The remaining will be hidden and can be expanded by the user.
+  Use Expandable to show only the first few items.
+  The remaining will be hidden and can be expanded by the user.
 */
 class Expandable extends Component {
   state = {
     expanded: false,
     toggleExpanded: () => {
       this.setState({
-        expanded: !this.state.expanded
+        expanded: !this.state.expanded,
       })
     },
     disabled: false,
     disable: () => {
       this.setState({
-        disabled: true
+        disabled: true,
       })
-    }
+    },
   }
 
   render() {
@@ -33,14 +34,12 @@ class Expandable extends Component {
 
 Expandable.propTypes = {
   expanded: PropTypes.bool,
-  disabled: PropTypes.bool
-};
+  disabled: PropTypes.bool,
+}
 
 Expandable.defaultProps = {
-  expanded: false
-};
+  expanded: false,
+}
 
 export default Expandable
-export {
-  ExpandableContext
-}
+export { ExpandableContext }
