+++ src/reusable/expandable/index.js
@@ -1,0 +1,6 @@
+import Expandable from './expandable'
+import ExpandableButton from './expandable-button'
+import ExpandableChildren from './expandable-children'
+import ExpandableProvider from './expandable-provider'
+
+export { Expandable, ExpandableButton, ExpandableChildren, ExpandableProvider }
