--- src/reusable/expandale/index.js
@@ -1,11 +1,0 @@
-import Expandable from './expandable';
-import ExpandableButton from './expandable-button';
-import ExpandableChildren from './expandable-children';
-import ExpandableProvider from './expandable-provider';
-
-export {
-  Expandable,
-  ExpandableButton,
-  ExpandableChildren,
-  ExpandableProvider
-}
